Deal with all warnings in runtime
runtime/application/src/OcpiServer.cxx
runtime/application/src/RemoteServer.cxx
runtime/container/src/ContainerApplication.cxx
runtime/container/src/ContainerArtifact.cxx
runtime/container/src/ContainerManager.cxx
runtime/container/src/ContainerPort.cxx
runtime/dataplane/rdma_drivers/dma/src/DtDmaXfer.cxx
runtime/dataplane/rdma_drivers/socket/src/DtSocketXfer.cxx
runtime/dataplane/transport/impl/src/OcpiPort.cxx
runtime/hdl-support/src/HdlSimServer.cxx
runtime/hdl-support/src/ocpihdl_main.cxx
runtime/hdl-support/src/testRpl_main.cxx
runtime/hdl/include/HdlDevice.h
runtime/hdl/src/HdlContainer.cxx
runtime/hdl/src/HdlDriver.cxx
runtime/hdl/src/HdlLSimDevice.cxx
runtime/hdl/src/HdlNetDriver.cxx
runtime/hdl/src/HdlWciControl.cxx
runtime/library/src/OcpiComponentLibrary.cxx
runtime/library/src/OcpiLibraryAssembly.cxx
runtime/library/src/OcpiLibraryManager.cxx
runtime/rcc/src/RccWorker.cxx
runtime/util/assembly/src/OcpiUtilAssembly.cxx
runtime/util/bin/src/ZipWriterDemo_main.cxx
runtime/util/ezxml/src/OcpiUtilEzxml.cxx
runtime/util/misc/include/OcpiUtilMisc.h
runtime/util/misc/src/OcpiExprEvaluator.cxx
runtime/util/misc/src/OcpiUtilMisc.cxx
runtime/util/property/src/OcpiUtilPValue.cxx
runtime/util/property/src/OcpiUtilProperty.cxx
runtime/util/property/src/OcpiUtilProtocol.cxx
runtime/util/zipfs/src/OcpiUtilZipFs.cxx

+++ runtime/hdl/include/HdlDevice.h
@@ -20,11 +20,11 @@ namespace OCPI {
     typedef uint32_t RomWord;
     static const unsigned
       ROM_NBYTES = 8*1024,
-      ROM_WIDTH  = sizeof(RomWord)*CHAR_BIT,
+      ROM_WIDTH  = (unsigned)sizeof(RomWord)*CHAR_BIT,
       ROM_WIDTH_BYTES = sizeof(RomWord),
       ROM_NWORDS = (ROM_NBYTES + ROM_WIDTH_BYTES-1) / ROM_WIDTH_BYTES,
       ROM_HEADER_WORDS = 4,
-      ROM_HEADER_BYTES = ROM_HEADER_WORDS * sizeof(RomWord);
+      ROM_HEADER_BYTES = ROM_HEADER_WORDS * (unsigned)sizeof(RomWord);
     class Device {
       HdlUUID m_UUID;
       OCPI::Util::Uuid m_loadedUUID;
